+++ CMakeLists.txt
@@ -36,6 +36,8 @@ add_executable(aaperturb
         include/fancy.h
         src/fancy.cpp
         include/threadpool.h
+        include/perturbrun.h
+        src/perturbrun.cpp
 )
 
 add_executable(single_perturb
@@ -49,6 +51,8 @@ add_executable(single_perturb
         src/molecules.cpp
         include/fancy.h
         src/fancy.cpp
+        include/perturbrun.h
+        src/perturbrun.cpp
 )
 
 add_executable(find_interfaces
